pyproject.toml
[project]
name = "democompany-identities"
version = "0.1.0"
description = "Generate corporate contractor identities and corporate emails."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
]

[project.scripts]
democompany-identities = "democompany_identities.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/democompany_identities"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "respx>=0.22.0",
    "ruff>=0.11.7",
]

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["src", "tests"]
extend-exclude = [".venv", "output", "logs"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]

[tool.pytest.ini_options]
addopts = "--cov=democompany_identities --cov-report=term-missing"
testpaths = ["tests"]
